fix: resolve all compilation errors from previous fixes
- bridge_auth.rs: Update Auth middleware to use EitherBody<BoxBody> for actix-web 4.13 compat
- bridge.rs: Fix SSE stream to use map() instead of filter_map() with correct error types
- bridge.rs: Import Emitter trait for notify endpoint
- bridge.rs: Fix allowed_headers to accept iterator
- commands.rs: Fix borrow lifetime in run_agent by separating state_ref from lock guard
- commands.rs: Re-export PipelineState from audio module
- audio/mod.rs: Re-export PipelineState for commands.rs access
- type_mode.rs: Use helper function instead of duplicate display_server import
- screen_router.rs: Cast scale_factor from f32 to f64

Verified: cargo check passes, 90 frontend tests pass across 12 files

+++ src-tauri/src/overlay/screen_router.rs
@@ -35,7 +35,7 @@ impl ScreenManager {
         {
             if let Ok(all) = xcap::Monitor::all() {
                 for m in &all {
-                    let scale_factor = m.scale_factor().unwrap_or(1.0);
+                    let scale_factor = m.scale_factor().unwrap_or(1.0) as f64;
                     monitors.push(MonitorInfo {
                         name: m.name().unwrap_or_else(|_| "unknown".into()),
                         x: m.x().unwrap_or(0),
